requirements.txt
torch
torchvision
Pillow
sklearn